requirements.txt
Flask
requests
beautifulsoup4
pandas
reportlab
openpyxl
gspread
google-auth-oauthlib
google-api-python-client
gunicorn